database/requirements.txt
# Database module dependencies for SentinelAI Audit Framework
supabase>=2.0.0
click>=8.0.0
python-dotenv>=0.19.0
pydantic>=1.10.0
psycopg2-binary>=2.9.0
